requirements-pc.txt
httpx==0.27.2
psutil==6.1.0
numpy==2.1.3
pydantic==2.9.2
safetensors==0.4.5
pyyaml>=6.0
tqdm>=4.66.0
torch
transformers
